Refactoring

+++ hashlab_listing/src/main/java/br/com/bruno96dantas/hashlab_listing/models/Product.java
@@ -19,7 +19,7 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Float price;
+    private double price;
     private String title;
     private String description;
 }
